Wallet_grouped to assist with verification
Wallets had been grouped and this is provided in the large datasheets, however I realised it may be easier for this processing to be done on my end. I have presented this data in a way which should be easy to handle. Note that half of the groups are related to scroll activity only (they should still correlate quite strongly on other chains). Multi chain assignments, obviously, will correlate strongly across all chains I reviewed.